requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --all-extras --annotation-style=line --output-file=requirements/dev.txt
#
argon2-cffi==23.1.0       # via django
argon2-cffi-bindings==21.2.0  # via argon2-cffi
asgiref==3.8.1            # via django
asttokens==3.0.0          # via stack-data
black==25.1.0             # via national-movie-database (setup.py)
brotli==1.1.0             # via whitenoise
certifi==2025.1.31        # via requests
cffi==1.17.1              # via argon2-cffi-bindings
charset-normalizer==3.4.1  # via requests
click==8.1.8              # via black
coverage[toml]==7.7.1     # via national-movie-database (setup.py), pytest-cov
crispy-bootstrap5==2024.10  # via national-movie-database (setup.py)
decorator==5.2.1          # via ipython
dj-database-url==2.3.0    # via national-movie-database (setup.py)
django[argon2]==5.1.7     # via crispy-bootstrap5, dj-database-url, django-appconf, django-crispy-forms, django-debug-toolbar, django-extensions, django-filter, django-js-asset, django-taggit, djangorestframework, model-bakery, national-movie-database (setup.py)
django-appconf==1.1.0     # via django-imagekit
django-crispy-forms==2.3  # via crispy-bootstrap5, national-movie-database (setup.py)
django-debug-toolbar==5.1.0  # via national-movie-database (setup.py)
django-extensions==3.2.3  # via national-movie-database (setup.py)
django-filter==25.1       # via national-movie-database (setup.py)
django-hashids==0.7.0     # via national-movie-database (setup.py)
django-imagekit==5.0.0    # via national-movie-database (setup.py)
django-js-asset==3.1.2    # via django-mptt
django-mptt==0.16.0       # via national-movie-database (setup.py)
django-registration-redux==2.13  # via national-movie-database (setup.py)
django-taggit==6.1.0      # via national-movie-database (setup.py)
djangorestframework==3.16.0  # via national-movie-database (setup.py)
environs==14.1.1          # via national-movie-database (setup.py)
executing==2.2.0          # via stack-data
faker==37.1.0             # via national-movie-database (setup.py)
graphviz==0.20.3          # via national-movie-database (setup.py)
gunicorn==23.0.0          # via national-movie-database (setup.py)
hashids==1.3.1            # via django-hashids
idna==3.10                # via requests
iniconfig==2.1.0          # via pytest
ipython==9.0.2            # via national-movie-database (setup.py)
ipython-pygments-lexers==1.1.1  # via ipython
isort==6.0.1              # via national-movie-database (setup.py)
jedi==0.19.2              # via ipython
marshmallow==3.26.1       # via environs
matplotlib-inline==0.1.7  # via ipython
mistune==3.1.3            # via national-movie-database (setup.py)
model-bakery==1.20.4      # via national-movie-database (setup.py)
mypy==1.15.0              # via national-movie-database (setup.py)
mypy-extensions==1.0.0    # via black, mypy
packaging==24.2           # via black, gunicorn, marshmallow, pytest
parso==0.8.4              # via jedi
pathspec==0.12.1          # via black
pexpect==4.9.0            # via ipython
pilkit==3.0               # via django-imagekit
pillow==11.1.0            # via national-movie-database (setup.py), pilkit
platformdirs==4.3.7       # via black
pluggy==1.5.0             # via pytest
prompt-toolkit==3.0.50    # via ipython
psycopg[c,pool]==3.2.6    # via national-movie-database (setup.py)
psycopg-c==3.2.6          # via psycopg
psycopg-pool==3.2.6       # via psycopg
ptyprocess==0.7.0         # via pexpect
pure-eval==0.2.3          # via stack-data
pycparser==2.22           # via cffi
pygments==2.19.1          # via ipython, ipython-pygments-lexers, national-movie-database (setup.py)
pytest==8.3.5             # via national-movie-database (setup.py), pytest-cov, pytest-django
pytest-cov==6.0.0         # via national-movie-database (setup.py)
pytest-django==4.10.0     # via national-movie-database (setup.py)
python-dotenv==1.1.0      # via environs
requests==2.32.3          # via stripe
ruff==0.11.2              # via national-movie-database (setup.py)
sqlparse==0.5.3           # via django, django-debug-toolbar
stack-data==0.6.3         # via ipython
stripe==11.6.0            # via national-movie-database (setup.py)
traitlets==5.14.3         # via ipython, matplotlib-inline
typing-extensions==4.13.0  # via dj-database-url, mypy, psycopg-pool, stripe
tzdata==2025.2            # via faker
urllib3==2.3.0            # via requests
wcwidth==0.2.13           # via prompt-toolkit
whitenoise[brotli]==6.9.0  # via national-movie-database (setup.py)
